line-service/requirements.txt
certifi==2018.11.29
chardet==3.0.4
Click==7.0
Flask==1.0.2
future==0.17.1
get==2018.11.19
idna==2.8
itsdangerous==1.1.0
Jinja2==2.10
line-bot-sdk==1.8.0
MarkupSafe==1.1.0
post==2018.11.20
public==2018.11.20
python-dotenv==0.10.1
query-string==2018.11.20
request==2018.11.20
requests==2.21.0
six==1.12.0
socketIO-client==0.7.2
urllib3==1.24.1
websocket-client==0.54.0
Werkzeug==0.14.1
